Cambiando de Float a Double Long y Lat de los aeropuertos

+++ src/main/java/com/flightspredictor/flights/domain/airports/dto/external/Location.java
@@ -1,7 +1,7 @@
 package com.flightspredictor.flights.domain.airports.dto.external;
 
 public record Location(
-        Float lat,
-        Float lon
+        Double lat,
+        Double lon
 ) {
 }
